Temp_Test/requirements.txt
google-cloud-vision==3.7.0
google-cloud-storage==2.16.0
python-dotenv==1.0.1
google-cloud-storage==2.16.0